Simplify code with test data reader
