+++ src/main/java/page/objects/BasePage.java
@@ -1,8 +1,7 @@
 package page.objects;
 
-import org.junit.jupiter.api.BeforeAll;
 import org.openqa.selenium.WebDriver;
-import test.data.TestDataReader;
+import test.data.TestData;
 import test.data.model.Address;
 import test.data.model.Credentials;
 import test.data.model.Url;
@@ -16,14 +15,14 @@ import java.nio.charset.StandardCharsets;
 import java.util.Properties;
 
 public class BasePage {
-    protected static TestDataReader testDataReader;
+    protected static TestData testDataReader;
     private final static String TEST_DATA_LOCATION = "src/configs/Configuration.properties";
     protected WebDriver driver;
 
     public static void loadConfig() throws IOException {
         Properties properties = new Properties();
         properties.load(new InputStreamReader(new FileInputStream(TEST_DATA_LOCATION), StandardCharsets.UTF_8));
-        testDataReader = new TestDataReader(new Cookie(properties), new Cookie(properties), new User(properties),
+        testDataReader = new TestData(new Cookie(properties), new Cookie(properties), new User(properties),
                 new User(properties), new User(properties), new User(properties), new Address(properties),
                 new Address(properties), new Address(properties), new Url(properties), new Url(properties),
                 new Url(properties), new Url(properties), new Url(properties), new Url(properties), new Url(properties),
